Changed hydraulics examples to match r3051 and r3053

+++ trunk/examples/hydraulics/checkvalve/system.cc
@@ -73,13 +73,13 @@ System::System(const string &name, bool bilateral, bool unilateral) : Group(name
     lCV->setMaximalContactForceLaw(new RegularizedUnilateralConstraint(new LinearRegularizedUnilateralConstraint(c, d)));
 
   ConstrainedNodeMec * n1 = new ConstrainedNodeMec("n1");
-  n1->setpFunction(new TabularFunction<double>(Vec("[0; .9; 1.1; 2.9; 3.1; 5]")*.1, "[4e5; 4e5; 2e5; 2e5; 4e5; 4e5]"));
+  n1->setpFunction(new TabularFunction<double(double)>(Vec("[0; .9; 1.1; 2.9; 3.1; 5]")*.1, "[4e5; 4e5; 2e5; 2e5; 4e5; 4e5]"));
   addLink(n1);
   n1->addOutFlow(lCV->getLine());
 
   ConstrainedNodeMec * n2 = new ConstrainedNodeMec("n2");
   addLink(n2);
-  n2->setpFunction(new ConstantFunction<double>(3e5));
+  n2->setpFunction(new ConstantFunction<double(double)>(3e5));
   n2->addInFlow(lCV->getLine());
 
 #ifdef HAVE_OPENMBVCPPINTERFACE
